fix bug with Perl 5.8.8
This took a while to figure out, but bigint that came bundled
with Perl 5.8.8 and earlier was not lexical!  To make matters
worse, the current version of Math-BigInt doesn't install on
Perl 5.8.8 so it isn't upgradable either.  To work around this
we will isolate the tests that use bigint in their own .t file
so that bigint doesn't muck up the float and double tests.